pyproject.toml
[tool.poetry]
name = "echostats"
version = "0.1.0"
description = "To calculate stats about EchoArena game"
authors = ["Laser Racer <cat@heaven.ist>"]
license = "GPLv3"
packages = [
    { include = "echostats", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
click = "^8.1.3"
pydantic = "^1.9.1"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
jupyter = "^1.0.0"
plotly = "^5.8.0"
dash = "^2.4.1"
jupyter-dash = "^0.4.2"
pandas = "^1.4.2"
ipywidgets = "^7.7.0"
jupyterlab = "^3.4.2"
mypy = "^0.961"

[tool.poetry.scripts]
echostats = "echostats:__main__"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
